+++ ontimize/components/o-form-data-component.class.ts
@@ -340,7 +340,7 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
     return this.defaultValue;
   }
 
-  public setValue(val: any, options?: IFormValueOptions, setDirty: boolean = false): void {
+  public setValue(val: any, options: IFormValueOptions = {}, setDirty: boolean = false): void {
     if (!PermissionsUtils.checkEnabledPermission(this.permissions)) {
       return;
     }
@@ -348,8 +348,10 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
       const newValue = val;
       const previousValue = this.oldValue;
       this.setFormValue(val, options, setDirty);
-      const changeType: number = (options && options.hasOwnProperty('changeType')) ? options.changeType : OValueChangeEvent.PROGRAMMATIC_CHANGE;
-      this.emitOnValueChange(changeType, newValue, previousValue);
+      if (options && options.emitModelToViewValueChange !== false) {
+        const changeType: number = (options && options.hasOwnProperty('changeType')) ? options.changeType : OValueChangeEvent.PROGRAMMATIC_CHANGE;
+        this.emitOnValueChange(changeType, newValue, previousValue);
+      }
     }
   }
 
@@ -515,7 +517,7 @@ export class OFormDataComponent extends OBaseComponent implements IFormDataCompo
   }
 
   get floatLabel(): FloatLabelType {
-    if(!this.labelVisible) {
+    if (!this.labelVisible) {
       this.floatLabel = 'never';
     }
     return this._floatLabel;