+++ ontimize/components/table/o-table.component.ts
@@ -17,7 +17,8 @@ import { CdkTableModule } from '@angular/cdk/table';
 import { CommonModule } from '@angular/common';
 import { DndModule } from '@churchs19/ng2-dnd';
 import { IOContextMenuContext } from '../contextmenu/o-context-menu.service';
-import { InputConverter } from '../../decorators';
+import { InputConverter, BooleanConverter } from '../../decorators/input-converter';
+import { NgxMaterialTimepickerModule } from 'ngx-material-timepicker';
 import { OContextMenuComponent } from '../contextmenu/o-context-menu-components';
 import { OContextMenuModule } from '../contextmenu/o-context-menu.module';
 import { OFormComponent } from '../form/o-form.component';
@@ -502,10 +503,9 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
   }
 
   protected _selectAllCheckboxVisible;
-  @InputConverter()
   set selectAllCheckboxVisible(value: boolean) {
-    this._selectAllCheckboxVisible = this.state['select-column-visible'] || value;
-    this.oTableOptions.selectColumn.visible = value;
+    this._selectAllCheckboxVisible = BooleanConverter(this.state['select-column-visible']) || BooleanConverter(value);
+    this.oTableOptions.selectColumn.visible = this._selectAllCheckboxVisible;
     this.updateSelectionColumnState();
   }
 
@@ -1558,15 +1558,18 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
 
     this.clearSelectionAndEditing();
     this.selectedRow(row);
-    this.editingCell = event.currentTarget;
+    if (event) {
+      this.editingCell = event.currentTarget;
+    }
     let rowData = {};
     this.keysArray.forEach((key) => {
       rowData[key] = row[key];
     });
     rowData[column.attr] = row[column.attr];
     this.editingRow = row;
-    column.editor.startEdition(rowData);
     column.editing = true;
+    column.editor.startEdition(rowData);
+    this.cd.detectChanges();
   }
 
   updateCellData(column: OColumn, data: any, saveChanges: boolean) {
@@ -1582,7 +1585,6 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
       Object.assign(this.editingRow, data);
     }
     this.editingRow = undefined;
-
     if (saveChanges && column.editor.updateRecordOnEdit) {
       let toUpdate = {};
       toUpdate[column.attr] = data[column.attr];
@@ -1622,7 +1624,6 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     } else if (this._oTableOptions.visibleColumns && !this._oTableOptions.selectColumn.visible && this._oTableOptions.visibleColumns[0] === OTableComponent.NAME_COLUMN_SELECT) {
       this._oTableOptions.visibleColumns.shift();
     }
-
   }
 
   public isAllSelected(): boolean {
@@ -2244,8 +2245,8 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     return result;
   }
 
-  getColumnInsertable(name):string{
-    return name+ this.getSuffixColumnInsertable();
+  getColumnInsertable(name): string {
+    return name + this.getSuffixColumnInsertable();
 
   }
 }
@@ -2271,7 +2272,8 @@ export class OTableComponent extends OServiceComponent implements OnInit, OnDest
     DndModule,
     OContextMenuModule,
     ObserversModule,
-    OMatSortModule
+    OMatSortModule,
+    NgxMaterialTimepickerModule
   ],
   exports: [
     OTableComponent,