+++ ontimize/util/codes.ts
@@ -67,12 +67,19 @@ export class Codes {
   public static COLUMN_TITLE_ALIGN_END = 'end';
   public static COLUMN_TITLE_ALIGN_AUTO = 'auto';
   public static AVAILABLE_COLUMN_TITLE_ALIGNS =
-  [Codes.COLUMN_TITLE_ALIGN_START, Codes.COLUMN_TITLE_ALIGN_CENTER, Codes.COLUMN_TITLE_ALIGN_END, Codes.COLUMN_TITLE_ALIGN_AUTO];
+    [Codes.COLUMN_TITLE_ALIGN_START, Codes.COLUMN_TITLE_ALIGN_CENTER, Codes.COLUMN_TITLE_ALIGN_END, Codes.COLUMN_TITLE_ALIGN_AUTO];
 
   public static O_MAT_ERROR_STANDARD = 'standard';
   public static O_MAT_ERROR_LITE = 'lite';
 
   public static O_INPUTS_OPTIONS_COLOR_ACCENT = 'accent';
+  public static HourFormat = {
+    TWELVE: 'hh:mm a',
+    TWENTY_FOUR: 'HH:mm a',
+  };
+
+  public static TWENTY_FOUR_HOUR_FORMAT = 24;
+  public static TWELVE_FOUR_HOUR_FORMAT = 12;
 
   static isDoubleClickMode(value: string): boolean {
     return Codes.DETAIL_MODE_DBLCLICK_VALUES.indexOf(value) !== -1;
@@ -86,5 +93,27 @@ export class Codes {
     let res = {};
     res[Codes.IS_DETAIL] = 'true';
     return res;
+  }
+
+  static formatString(format: number) {
+    return (format === Codes.TWENTY_FOUR_HOUR_FORMAT ? Codes.HourFormat.TWENTY_FOUR : Codes.HourFormat.TWELVE);
+  }
+  
+
+  static isHourInputAllowed(e: KeyboardEvent): boolean {
+    // Allow: backspace, delete, tab, escape, enter
+    if ([46, 8, 9, 27, 13].some(n => n === e.keyCode) ||
+      (e.key === ':') ||
+      // Allow: Ctrl/cmd+A
+      (e.keyCode === 65 && (e.ctrlKey === true || e.metaKey === true)) ||
+      // Allow: Ctrl/cmd+C
+      (e.keyCode === 67 && (e.ctrlKey === true || e.metaKey === true)) ||
+      // Allow: Ctrl/cmd+X
+      (e.keyCode === 88 && (e.ctrlKey === true || e.metaKey === true)) ||
+      // Allow: home, end, left, right, up, down
+      (e.keyCode >= 35 && e.keyCode <= 40)) {
+      return true;
+    }
+    return !((e.keyCode < 48 || e.keyCode > 57) && (e.keyCode < 96 || e.keyCode > 105));
   }
 }